Fix all warnings on win32 build
svn:r2127

+++ src/or/rephist.c
@@ -468,9 +468,9 @@ int rep_hist_bandwidth_assess(time_t when) {
   r = find_largest_max(read_array);
   w = find_largest_max(write_array);
   if (r>w)
-    return w/(double)NUM_SECS_ROLLING_MEASURE;
+    return (int)(w/(double)NUM_SECS_ROLLING_MEASURE);
   else
-    return r/(double)NUM_SECS_ROLLING_MEASURE;
+    return (int)(r/(double)NUM_SECS_ROLLING_MEASURE);
 
   return 0;
 }
